Migrate all creep_* to role_*
