+++ src/config_creep_routing.js
@@ -144,7 +144,7 @@ Creep.prototype.initRouting = function() {
   if (this.memory.role == 'harvester') {
     targetId = 'harvester';
   }
-  if (this.memory.role == 'builder') {
+  if (this.memory.role == 'upgrader') {
     targetId = this.room.controller.id;
   }
 
@@ -248,7 +248,10 @@ Creep.prototype.followPath = function(action) {
   // return false;
   // }
 
-  let unit = require('creep_' + this.memory.role);
+  let unit = roles[this.memory.role];
+  if (!unit) {
+    unit = require('creep_' + this.memory.role);
+  }
   if (!this.memory.routing.targetId && routePos == route.length - 1) {
     if (unit.getTargetId) {
       this.memory.routing.targetId = unit.getTargetId(this);
@@ -263,7 +266,10 @@ Creep.prototype.followPath = function(action) {
 };
 
 Creep.prototype.moveByPathMy = function(route, routePos, start, target, skipPreMove, action) {
-  let unit = require('creep_' + this.memory.role);
+  let unit = roles[this.memory.role];
+  if (!unit) {
+    unit = require('creep_' + this.memory.role);
+  }
   // Somehow reset the pathPos if the path has changed?!
   let path = this.room.getPath(route, routePos, start, target);
   if (!path) {
