+++ src/require.js
@@ -9,6 +9,7 @@
   require('config_creep_resources');
   require('config_creep_harvest');
   require('config_creep_mineral');
+  require('config_creep_nextroom');
   require('config_creep_routing');
   require('config_roomPosition_structures');
   require('config_room');
@@ -24,7 +25,31 @@
   require('config_room_routing');
   require('config_room_wallsetter');
   require('config_string');
+  require('role_atkeeper');
+  require('role_atkeepermelee');
+  require('role_autoattackmelee');
+  require('role_carry');
+  require('role_defender');
+  require('role_defendmelee');
+  require('role_defendranged');
+  require('role_extractor');
+  require('role_harvester');
+  require('role_mineral');
+  require('role_nextroomer');
+  require('role_nextroomerattack');
+  require('role_planer');
+  require('role_powerattacker');
+  require('role_powerdefender');
+  require('role_powerhealer');
+  require('role_powertransporter');
+  require('role_reserver');
+  require('role_repairer');
+  require('role_scout');
+  require('role_scoutnextroom');
+  require('role_sourcer');
+  require('role_squadsiege');
+  require('role_squadheal');
   require('role_storagefiller');
   require('role_structurer');
   require('role_towerfiller');
-  require('role_scout');
+  require('role_upgrader');
