+++ src/role_defendranged.js
@@ -1,35 +1,27 @@
 'use strict';
 
-module.exports.get_part_config = function(room, energy, heal) {
+/*
+ * defendranged is called after 'threshold' when a room is attacked
+ * 
+ * Tries to fight against the hostile creeps from ramparts if possible
+ */
+
+roles.defendranged = {};
+
+roles.defendranged.getPartConfig = function(room, energy, heal) {
   var parts = [MOVE, RANGED_ATTACK, MOVE, RANGED_ATTACK, MOVE, RANGED_ATTACK];
   return room.get_part_config(energy, parts);
 };
 
-module.exports.energyRequired = function(room) {
+roles.defendranged.get_part_config = roles.defendranged.getPartConfig;
+
+roles.defendranged.energyRequired = function(room) {
   return Math.max(200, room.energyAvailable);
 };
 
-module.exports.energyBuild = function(room) {
+roles.defendranged.energyBuild = function(room) {
   return Math.max(200, room.energyAvailable);
 };
-
-function recycleCreep(creep) {
-  creep.say('recycle');
-  if (creep.room.controller && creep.room.controller.my) {
-    if (creep.memory.countdown > 0) {
-      creep.memory.countdown -= 1;
-      creep.say('rnd');
-      creep.moveRandom();
-      return false;
-    }
-  }
-  if (creep.room.name != creep.memory.base) {
-    if (creep.stayInRoom()) {
-      return true;
-    }
-  }
-  return Creep.recycleCreep(creep);
-}
 
 // TODO This overwrites the target so redo and enable again
 //module.exports.action = function(creep) {
@@ -60,8 +52,26 @@ function recycleCreep(creep) {
 //  return creep.fightRanged(target);
 //};
 
-module.exports.execute = function(creep) {
+roles.defendranged.execute = function(creep) {
   creep.memory.countdown = creep.memory.countdown || 100;
+
+  let recycleCreep = function(creep) {
+    creep.say('recycle');
+    if (creep.room.controller && creep.room.controller.my) {
+      if (creep.memory.countdown > 0) {
+        creep.memory.countdown -= 1;
+        creep.say('rnd');
+        creep.moveRandom();
+        return false;
+      }
+    }
+    if (creep.room.name != creep.memory.base) {
+      if (creep.stayInRoom()) {
+        return true;
+      }
+    }
+    return Creep.recycleCreep(creep);
+  };
 
   let hostiles = creep.room.find(FIND_HOSTILE_CREEPS, {
     filter: creep.room.findAttackCreeps
