+++ src/role_mineral.js
@@ -1,0 +1,30 @@
+'use strict';
+
+/*
+ * mineral manages the mineral distributions
+ * 
+ * Checks for room reactions and transfers mineral to the associated labs
+ * Checks for boost request and transfers mineral to the associated labs
+ * Fills the labs with energy
+ */
+
+roles.mineral = {};
+roles.mineral.getPartConfig = function(room, energy, heal) {
+  var parts = [MOVE, CARRY];
+  return room.get_part_config(energy, parts);
+};
+roles.mineral.get_part_config = roles.mineral.getPartConfig;
+
+roles.mineral.energyBuild = function(room, energy, source, heal) {
+  var max = 1000;
+  energy = Math.max(250, Math.min(max, room.energyCapacityAvailable));
+  return energy;
+};
+
+roles.mineral.action = function(creep) {
+  return creep.handleMineralCreep();
+};
+
+roles.mineral.execute = function(creep) {
+  return creep.handleMineralCreep();
+};
