+++ src/role_planer.js
@@ -1,17 +1,26 @@
 'use strict';
 
-module.exports.stayInRoom = true;
+/*
+ * planer builds up construction sites
+ * 
+ * Moves to the construction sites, does random walk to prevent traffic jam
+ * builds up the structure
+ */
 
-module.exports.get_part_config = function(room, energy, heal) {
+
+roles.planer = {};
+roles.planer.stayInRoom = true;
+
+roles.planer.get_part_config = function(room, energy, heal) {
   var parts = [MOVE, CARRY, MOVE, WORK];
   return room.get_part_config(energy, parts);
 };
 
-module.exports.energyBuild = function(room, energy) {
+roles.planer.energyBuild = function(room, energy) {
   return Math.max(250, Math.min(energy, 3000));
 };
 
-module.exports.action = function(creep) {
+roles.planer.action = function(creep) {
   var methods = [Creep.getEnergy];
   if (creep.room.storage && creep.room.storage.store.energy > config.creep.energyFromStorageThreshold) {
     methods = [Creep.getEnergyFromStorage];
@@ -30,7 +39,7 @@ module.exports.action = function(creep) {
 };
 
 
-module.exports.execute = function(creep) {
+roles.planer.execute = function(creep) {
   var methods = [Creep.getEnergy];
   if (creep.room.storage && creep.room.storage.store.energy > config.creep.energyFromStorageThreshold) {
     methods = [Creep.getEnergyFromStorage];
