+++ src/role_sourcer.js
@@ -1,17 +1,32 @@
 'use strict';
 
-module.exports.buildRoad = true;
-module.exports.killPrevious = true;
+/*
+ * Harvesting sources is done by sourcer
+ * 
+ * Moves to the source and gets energy 
+ * In external rooms builds a container
+ * In internal rooms transfers to the link
+ * 
+ * If 'threshold' energy is in the container or on the ground
+ * a carry is called
+ */
+
+roles.sourcer = {};
+
+roles.sourcer.buildRoad = true;
+roles.sourcer.killPrevious = true;
 
 // TODO should be true, but flee must be fixed before 2016-10-13
-module.exports.flee = false;
+roles.sourcer.flee = false;
 
-module.exports.get_part_config = function(room, energy, heal) {
+roles.sourcer.getPartConfig = function(room, energy, heal) {
   var parts = [MOVE, CARRY, WORK, WORK, WORK, WORK, MOVE, MOVE, WORK, HEAL, MOVE, HEAL, MOVE, WORK, WORK, WORK, WORK, WORK, MOVE, MOVE, MOVE, MOVE, MOVE];
   return room.get_part_config(energy, parts);
 };
 
-module.exports.preMove = function(creep, directions) {
+roles.sourcer.get_part_config = roles.sourcer.getPartConfig;
+
+roles.sourcer.preMove = function(creep, directions) {
   // Misplaced spawn
   if (creep.room.memory.misplacedSpawn || creep.room.controller.level < 3) {
     creep.say('mis', true);
@@ -29,7 +44,6 @@ module.exports.preMove = function(creep, directions) {
       return true;
     }
   }
-
 
   if (!creep.room.controller) {
     var target = creep.pos.findClosestByRange(FIND_HOSTILE_CREEPS, {
@@ -83,7 +97,7 @@ module.exports.preMove = function(creep, directions) {
   }
 };
 
-module.exports.energyBuild = function(room, energy, source, heal) {
+roles.sourcer.energyBuild = function(room, energy, source, heal) {
   var max = 700;
   // TODO Only three parts for external sourcer (Double check how many parts)
   //  room.log('creep_sourcer.energyBuild source: ' + JSON.stringify(source));
@@ -94,12 +108,12 @@ module.exports.energyBuild = function(room, energy, source, heal) {
   return energy;
 };
 
-module.exports.died = function(name, memory) {
+roles.sourcer.died = function(name, memory) {
   console.log(name, 'died', memory.base, memory.source.roomName);
   delete Memory.creeps[name];
 };
 
-module.exports.action = function(creep) {
+roles.sourcer.action = function(creep) {
   // TODO Fix for sourcers without routing.targetId
   if (!creep.memory.routing.targetId) {
     if (!Game.rooms[creep.memory.source.roomName]) {
@@ -140,148 +154,13 @@ module.exports.action = function(creep) {
     }
   }
 
-  work(creep);
+  creep.handleSourcer();
   return true;
 };
 
-function checkContainer(creep) {
-  if (creep.room.name == creep.memory.base) {
-    return false;
-  }
-  // TODO Not in base room
-  var objects = creep.pos.findInRange(FIND_STRUCTURES, 0, {
-    filter: function(object) {
-      if (object.structureType == STRUCTURE_CONTAINER) {
-        return true;
-      }
-      return false;
-    }
-  });
-  if (objects.length === 0) {
-    if (creep.carry.energy >= 50) {
-      let constructionSites = creep.pos.findInRange(FIND_CONSTRUCTION_SITES, 0, {
-        filter: function(object) {
-          if (object.structureType != STRUCTURE_CONTAINER) {
-            return false;
-          }
-          return true;
-        }
-      });
-      if (constructionSites.length > 0) {
-        let returnCode = creep.build(constructionSites[0]);
-        if (returnCode != OK) {
-          creep.log('build container: ' + returnCode);
-        }
-        return true;
-      }
-
-      let returnCode = creep.pos.createConstructionSite(STRUCTURE_CONTAINER);
-      if (returnCode == OK) {
-        creep.log('Create cs for container');
-        return true;
-      }
-      if (returnCode == ERR_INVALID_TARGET) {
-        let constructionSites = creep.pos.findInRange(FIND_CONSTRUCTION_SITES, 0);
-        for (let constructionSite of constructionSites) {
-          constructionSite.remove();
-        }
-        return false;
-      }
-      if (returnCode != ERR_FULL) {
-        creep.log('Container: ' + returnCode + ' pos: ' + creep.pos);
-      }
-      return false;
-    }
-  }
-  if (objects.length > 0) {
-    let object = objects[0];
-    if (object.hits < object.hitsMax) {
-      creep.repair(object);
-    }
-  }
-}
-
-function killPreviousSourcerer(creep) {
-  var other_sourcer = creep.pos.findClosestByRange(FIND_MY_CREEPS, {
-    filter: function(object) {
-      if (object.id == creep.id) {
-        return false;
-      }
-      if (object.memory.role == 'sourcer' && object.memory.target_id == creep.memory.target_id) {
-        return true;
-      }
-      return false;
-    }
-  });
-  if (other_sourcer === null) {
-    return false;
-  }
-
-  var range = creep.pos.getRangeTo(other_sourcer);
-  if (range == 1) {
-    if (other_sourcer.ticksToLive > creep.ticksToLive) {
-      creep.log('sourcer: Kill me');
-      creep.suicide();
-    } else {
-      //creep.log('Kill other sourcerer ' + range + ' ' + creep.name + ' ' + creep.memory.target_id);
-      other_sourcer.memory.killed = true;
-      creep.log('sourcer: Kill other');
-      other_sourcer.suicide();
-    }
-  }
-}
-
-function work(creep) {
-  creep.setNextSpawn();
-  creep.spawnReplacement();
-  let room = Game.rooms[creep.room.name];
-
-  // TODO legacy stuff
-  let targetId = creep.memory.target_id;
-  if (creep.memory.routing) {
-    targetId = creep.memory.routing.targetId;
-  } else {
-    console.log('No routing');
-  }
-
-  var source = Game.getObjectById(targetId);
-
-  let target = source;
-  let returnCode = creep.harvest(source);
-  if (returnCode != OK && returnCode != ERR_NOT_ENOUGH_RESOURCES) {
-    creep.log('harvest: ' + returnCode);
-    return false;
-  }
-
-  checkContainer(creep);
-
-  if (!creep.room.controller || !creep.room.controller.my || creep.room.controller.level >= 2) {
-    creep.spawnCarry();
-  }
-
-  if (creep.room.name == creep.memory.base) {
-    if (!creep.memory.link) {
-      let links = creep.pos.findInRange(FIND_MY_STRUCTURES, 1, {
-        filter: function(object) {
-          if (object.structureType == STRUCTURE_LINK) {
-            return true;
-          }
-          return false;
-        }
-      });
-      if (links.length > 0) {
-        creep.memory.link = links[0].id;
-      }
-    }
-
-    let link = Game.getObjectById(creep.memory.link);
-    creep.transfer(link, RESOURCE_ENERGY);
-  }
-}
-
-module.exports.execute = function(creep) {
+roles.sourcer.execute = function(creep) {
   creep.log('Execute!!!');
   creep.memory.routing.targetReached = true;
-  work(creep);
+  creep.handleSourcer();
   //  throw new Error();
 };
