+++ src/role_towerfiller.js
@@ -1,5 +1,11 @@
 'use strict';
 
+/*
+ * towerfiller is called when the room is under attack
+ * 
+ * Moves to the associated tower and transfers the energy from the link to the tower
+ */
+
 roles.towerfiller = {};
 roles.towerfiller.energyBuild = function(room, energy) {
   return 200;
